T9044 Move `getTranslatedText()` to `/utils`
